python_pytest/requirements.txt
pytest
requests
allure-pytest
selenium
python-benedict
deepdiff
requests